backend/requirements.txt
flask
flask-cors
pypdf
openai
python-dotenv
pydantic
httpx==0.27.2

